+++ src/main/java/com/naver/booking/api/dao/ProductApiDao.java
@@ -24,8 +24,8 @@ public class ProductApiDao {
 		this.jdbc = new NamedParameterJdbcTemplate(dataSource);
 	}
 	
-	public List<ProductApiDto> SelectProductsApi(int categoryId, int start){
-		Map<String, Integer> params = new HashMap<>();
+	public List<ProductApiDto> SelectProductsApi(Long categoryId, Long start){
+		Map<String, Long> params = new HashMap<>();
 		params.put("start", start);
 		if(categoryId == 0) {
 			return jdbc.query(ProductApiDaoSqls.SELECT_ALL_PRODUCTS_API, params, rowMapper);			
@@ -34,12 +34,12 @@ public class ProductApiDao {
 		return jdbc.query(ProductApiDaoSqls.SELECT_PRODUCTS_API, params, rowMapper);
 	}
 	
-	public int SelectProductCountByCategory(int categoryId) {
+	public Long SelectProductCountByCategory(Long categoryId) {
 		if(categoryId == 0) {
-			return jdbc.queryForObject(ProductApiDaoSqls.SELECT_ALL_PRODUCT_COUNT_BY_CATEGORY, Collections.emptyMap(), Integer.class);
+			return jdbc.queryForObject(ProductApiDaoSqls.SELECT_ALL_PRODUCT_COUNT_BY_CATEGORY, Collections.emptyMap(), Long.class);
 		}
-		Map<String, Integer> params = Collections.singletonMap("categoryId", categoryId);
-		return jdbc.queryForObject(ProductApiDaoSqls.SELECT_PRODUCT_COUNT_BY_CATEGORY, params, Integer.class);
+		Map<String, Long> params = Collections.singletonMap("categoryId", categoryId);
+		return jdbc.queryForObject(ProductApiDaoSqls.SELECT_PRODUCT_COUNT_BY_CATEGORY, params, Long.class);
 		
 	}
 }
